// src/r/types.ts
export namespace SharedNS {
  export interface IOptions {
    cwd?: string;
  }
}

/**
 * Release namespace
 */
export namespace ReleaseNS {
  export interface ILernaConfig {
    version: string;
    [key: string]: unknown;
  }

  export interface IModifyReleaseArgumentsContext {
    arguments: string[];
    version: string;
    tag: string;
  }

  export interface IReadyContext {
    cwd: string;
  }

  export interface IReleasedContext {
    cwd: string;
    version: string;
    tag: string;
  }

  export interface IPromptAnswers {
    bump: string;
    customVersion: string;
    npmTag: string;
    customNpmTag: string;
  }

  export interface IOptions extends SharedNS.IOptions {
    /**
     * Generate changelog or not.
     */
    changelog?: boolean;
    /**
     * Run in band or not.
     */
    runInBand?: boolean;
    /**
     * Ignore scripts, it will ignore the npm life cycle hooks (e.g. prepublishOnly).
     */
    ignoreScripts?: boolean;
    /**
     * Execute custom build script before release.
     *
     * It is recommended to run topological build before actual publishing process,
     * and then ignore the build script (with `ignoreScripts`) of each package during
     * actual publishing release.
     */
    build?: boolean | string;
    ready?: (ctx: IReadyContext) => Promise<unknown>;
    modifyReleaseArguments?: (ctx: IModifyReleaseArgumentsContext) => string[];
    released?: (ctx: IReleasedContext) => Promise<unknown>;
  }
}